alais

+++ 0x08-recursion/4-pow_recursion.c
@@ -7,7 +7,7 @@
  * Return: ll
  */
 
-int _pow_recursion(long x, long y)
+long _pow_recursion(int x, int y)
 {
 	if (y < 0)
 		return (-1);
